Fixed Font Renderer, cleaned up GL code, removed AWT, added Gradient
