--- src/main/java/game/platform/FillStyle.java
@@ -1,5 +1,0 @@
-package game.platform;
-
-public interface FillStyle {
-
-}
