+++ src/main/java/game/platform/Paint.java
@@ -1,0 +1,5 @@
+package game.platform;
+
+public interface Paint {
+    void apply(Renderer renderer, int x, int y, int width, int height);
+}
