+++ src/main/java/game/platform/Transform.java
@@ -1,8 +1,66 @@
 package game.platform;
 
-public interface Transform {
-    Transform copy();
+import lombok.Getter;
+import org.joml.Matrix4f;
+import org.joml.Vector2f;
+import org.joml.Vector3f;
 
-    Transform translate(int x, int y);
-    Transform inverse();
+@Getter
+public class Transform {
+    private final Matrix4f matrix;
+
+    public Transform() {
+        this.matrix = new Matrix4f();
+        this.matrix.identity();
+    }
+
+    public Transform(Matrix4f matrix) {
+        this.matrix = matrix;
+    }
+
+    public static Transform fromScreenSpace(int width, int height) {
+        Transform transform = new Transform();
+        transform.translate(-1.0f, 1.0f);  // Change Y from -1 to 1
+        transform.scale(2.0f / width, -2.0f / height);  // Negate Y scale
+        return transform;
+    }
+
+    public static Transform createTranslate(float x, float y) {
+        Transform transform = new Transform();
+        transform.translate(x, y);
+        return transform;
+    }
+
+    public Transform translate(Number x, Number y) {
+        matrix.translate(x.floatValue(), y.floatValue(), 0);
+        return this;
+    }
+
+    public Transform scale(Number v, Number v1) {
+        matrix.scale(v.floatValue(), v1.floatValue(), 1);
+        return this;
+    }
+
+    public Transform compose(Transform other) {
+        Matrix4f result = new Matrix4f();
+        matrix.mul(other.matrix, result);
+        return new Transform(result);
+    }
+
+    public Vector2f apply(Vector2f point) {
+        Vector3f result = new Vector3f();
+        // matrix * point
+        matrix.transformPosition(point.x, point.y, 0, result);
+        return new Vector2f(result.x, result.y);
+    }
+
+    public Transform copy() {
+        return new Transform(new Matrix4f(matrix));
+    }
+
+    public Transform inverse() {
+        Matrix4f inverseMatrix = new Matrix4f();
+        matrix.invert(inverseMatrix);
+        return new Transform(inverseMatrix);
+    }
 }
