+++ src/main/java/game/transition/CompositeTransition.java
@@ -2,8 +2,6 @@ package game.transition;
 
 import game.platform.Renderer;
 
-import java.awt.*;
-import java.awt.image.BufferedImage;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
