+++ src/main/java/game/transition/TransitionFactory.java
@@ -2,8 +2,6 @@ package game.transition;
 
 import game.platform.FrameBuffer;
 
-import java.awt.image.BufferedImage;
-
 @FunctionalInterface
 public interface TransitionFactory {
     Transition create(FrameBuffer source, FrameBuffer target, Runnable onComplete);
